PyBar V0.5 - Serando telas entre arquivos.
