+++ Clientes.h
@@ -1,0 +1,9 @@
+
+// Assinatura das Funcoes do menu cliente
+
+void menu_cliente(void);
+void tela_pedidos(void);
+void tela_comidas(void);
+void tela_bebidas(void);
+void tela_visualizar_pedido(void);
+void tela_ajuda(void);
