+++ Estoque.h
@@ -1,0 +1,10 @@
+
+// Assinatura das Funcoes do menu estoque
+
+void tela_estoque(void);
+void tela_cadastro_item(void);
+void tela_procura_item(void);
+void tela_lista_itens(void);
+void tela_excluir_item(void);
+void tela_atualizar_item(void);
+void tela_atualizacao(void);
