+++ Gerencia.h
@@ -1,0 +1,11 @@
+
+// Assinatura das Funcoes do menu gerencia
+
+void acesso_gerencia(void);
+void menu_cadastramento(void);
+void menu_login(void);
+void menu_gerencia(void);
+void tela_pedidos_gerencia(void);
+void tela_caixa(void);
+void tela_relatorio(void);
+void tela_funcionarios(void);
